+++ includes/model/tournament.php
@@ -22,6 +22,7 @@ class Ekc_Tournament implements JsonSerializable {
 	private $swiss_system_additional_rounds;
 	private $swiss_system_slide_match_rounds;
 	private $swiss_system_round_time;
+	private $swiss_system_start_pitch;
 	private $shareable_link_email_text;
 	private $shareable_link_url_prefix;
 	private $shareable_link_sender_email;
@@ -160,6 +161,14 @@ class Ekc_Tournament implements JsonSerializable {
 
 	public function set_swiss_system_round_time(?int $swiss_system_round_time) {
 		$this->swiss_system_round_time = $swiss_system_round_time;
+	}
+
+	public function get_swiss_system_start_pitch() {
+		return $this->swiss_system_start_pitch;
+	}
+
+	public function set_swiss_system_start_pitch(?int $swiss_system_start_pitch) {
+		$this->swiss_system_start_pitch = $swiss_system_start_pitch;
 	}
 
 	public function get_shareable_link_email_text() {